仓库提交

+++ ruoyi-modules/ruoyi-modules-goods/src/main/java/com/ruoyi/goods/service/impl/TWarehouseServiceImpl.java
@@ -1,6 +1,8 @@
 package com.ruoyi.goods.service.impl;
 
 import java.util.List;
+
+import com.ruoyi.common.core.exception.ServiceException;
 import com.ruoyi.common.core.utils.DateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -41,6 +43,7 @@ public class TWarehouseServiceImpl implements ITWarehouseService
     @Override
     public List<TWarehouse> selectTWarehouseList(TWarehouse tWarehouse)
     {
+        tWarehouse.setParentId(0L);
         return tWarehouseMapper.selectTWarehouseList(tWarehouse);
     }
 
@@ -53,6 +56,7 @@ public class TWarehouseServiceImpl implements ITWarehouseService
     @Override
     public int insertTWarehouse(TWarehouse tWarehouse)
     {
+        tWarehouse.setWarehouseType(1);
         tWarehouse.setCreateTime(DateUtils.getNowDate());
         return tWarehouseMapper.insertTWarehouse(tWarehouse);
     }
@@ -66,6 +70,11 @@ public class TWarehouseServiceImpl implements ITWarehouseService
     @Override
     public int updateTWarehouse(TWarehouse tWarehouse)
     {
+        TWarehouse house = new TWarehouse();
+        house.setSysUserId(tWarehouse.getSysUserId());
+        house.setAddress(tWarehouse.getAddress());
+        house.setPhone(tWarehouse.getPhone());
+        tWarehouseMapper.updateTWarehouseParent(house);
         tWarehouse.setUpdateTime(DateUtils.getNowDate());
         return tWarehouseMapper.updateTWarehouse(tWarehouse);
     }
@@ -91,6 +100,38 @@ public class TWarehouseServiceImpl implements ITWarehouseService
     @Override
     public int deleteTWarehouseById(Long id)
     {
+        List<TWarehouse> list = tWarehouseMapper.selectTWarehouseByParentId(id);
+        if(list != null && list.size()> 0){
+            throw new ServiceException("有所属子仓库不能删除");
+        }
         return tWarehouseMapper.deleteTWarehouseById(id);
+    }
+
+    @Override
+    public int insertTWarehouseChild(TWarehouse tWarehouse) {
+        if(tWarehouse.getParentId() == null){
+            throw new ServiceException("父id不能为空");
+        }
+        TWarehouse house = tWarehouseMapper.selectTWarehouseById(tWarehouse.getParentId());
+        if(house == null){
+            throw new ServiceException("父仓库不存在");
+        }
+        tWarehouse.setWarehouseType(2);
+        tWarehouse.setAddress(house.getAddress());
+        tWarehouse.setSysUserId(house.getSysUserId());
+        tWarehouse.setPhone(house.getPhone());
+        tWarehouse.setCreateTime(DateUtils.getNowDate());
+        return tWarehouseMapper.insertTWarehouse(tWarehouse);
+    }
+
+    @Override
+    public int updateTWarehouseChild(TWarehouse tWarehouse) {
+        tWarehouse.setUpdateTime(DateUtils.getNowDate());
+        return tWarehouseMapper.updateTWarehouse(tWarehouse);
+    }
+
+    @Override
+    public List<TWarehouse> selectTWarehouseListChild(TWarehouse tWarehouse) {
+        return tWarehouseMapper.selectTWarehouseByParentId(tWarehouse.getId());
     }
 }
